Support inline links with proper symbol counting
Inline links in vk format are expressions like [urlOrId|text]. While
they can be trivially matched with regex and replaced with <a href>,
there is a caveat that we need to stay within telegram message length
limit by splitting messages, and to avoid splitting them without
neccesity we have to count symbols excluding the URLs in inline links.
This commit implements such logic and replaces vk links with html tags
which will be parsed by telegram. Here we also move html.EscapeString
call to be after we counted rendered characters, which should guarantee
that we count everything properly and only split messages when their
would-be rendered length is in fact above limit.