Improve handling of repost chains
Before that commit if we had some page repost from another page and add
a comment, the comment was sent to telegram as a separate message
instead of being a reply. It gets fixed in this commit with the
following logic:

- each VK post may be split to several tg messages due to either long text or
  different media types, or both
- when a post is split because of above reasons, sendText or sendWithAttachments
  function returns the first message in a chain
- when crossposting a chain of reposts, the returned first message will
  be used as a reply-to message for higher level posts.

There is also another small change here: even if the user configured a
subscription to omit the reference to source, the reference will be
added when the source reposts something from external pages. I think it
is quite reasonable behavior and makes more clear that the message is a
repost and where it came from.